Break up partition tests
