+++ prechi/test/test_partition.h
@@ -3,4 +3,12 @@
 
 #include "../src/prechi_partition.h"
 
+typedef struct {
+  float *weights;
+  int *counts;
+} TestData;
+
+TestData *create_test_data(int count);
+void destroy_test_data(TestData *td);
+
 #endif
